fixes #1407 :: 64-bit breakage (Stdout dependencies)
Thanks to fvbommel

+++ tango/io/stream/Format.d
@@ -257,7 +257,7 @@ class FormatOutput(T) : OutputFilter
 
         **********************************************************************/
 
-        private final size_t sink (T[] s)
+        private final uint sink (T[] s)
         {
                 return output.write (s);
         }
